# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Family Finance & Chores API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /families/{familyId}/chores:
    get:
      operationId: listChores
      tags: [chores]
      summary: List all chores for a family
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chore"
    post:
      operationId: createChore
      tags: [chores]
      summary: Create a new chore
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChoreInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
  /families/{familyId}/chores/{choreId}:
    get:
      operationId: getChore
      tags: [chores]
      summary: Get chore details
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: integer
        - name: choreId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
    patch:
      operationId: updateChore
      tags: [chores]
      summary: Update chore (status, details)
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: integer
        - name: choreId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChoreUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
    delete:
      operationId: deleteChore
      tags: [chores]
      summary: Delete a chore
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: integer
        - name: choreId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Coin Transactions ──────────────────────────────────────────────────────
...
  /families/{familyId}/allowances:
    get:
      operationId: listAllowances
      tags: [allowances]
      summary: List allowances for a family
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Allowance"
    post:
      operationId: createAllowance
      tags: [allowances]
      summary: Create a recurring allowance for a member
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllowanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allowance"
  /families/{familyId}/allowances/{allowanceId}:
    patch:
      operationId: updateAllowance
      tags: [allowances]
      summary: Update an allowance
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: integer
        - name: allowanceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllowanceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allowance"
    delete:
      operationId: deleteAllowance
      tags: [allowances]
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: integer
        - name: allowanceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Receipt Scanning ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    Chore:
      type: object
      required: [id, familyId, title, coinReward, status, createdAt]
      properties:
        id:
          type: integer
        familyId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        coinReward:
          type: integer
        status:
          type: string
          enum: [pending, completed, approved, rejected]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        isRecurring:
          type: boolean
        recurringFrequency:
          type: ["string", "null"]
          enum: [daily, weekly, monthly, null]
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ChoreInput:
      type: object
      required: [title, coinReward]
      properties:
        title:
          type: string
        description:
          type: string
        coinReward:
          type: integer
        assigneeId:
          type: integer
        dueDate:
          type: string
          format: date
        isRecurring:
          type: boolean
        recurringFrequency:
          type: string
          enum: [daily, weekly, monthly]
    ChoreUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        coinReward:
          type: integer
        status:
          type: string
          enum: [pending, completed, approved, rejected]
        assigneeId:
          type: integer
        dueDate:
          type: string
          format: date

    # ── Transaction ──────────────────────────────────────────────────────────
...
    Allowance:
      type: object
      required: [id, familyId, memberId, memberName, coinAmount, frequency, isActive, createdAt]
      properties:
        id:
          type: integer
        familyId:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        coinAmount:
          type: integer
        frequency:
          type: string
          enum: [daily, weekly, monthly]
        isActive:
          type: boolean
        nextPayDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    AllowanceInput:
      type: object
      required: [memberId, coinAmount, frequency]
      properties:
        memberId:
          type: integer
        coinAmount:
          type: integer
        frequency:
          type: string
          enum: [daily, weekly, monthly]
    AllowanceUpdate:
      type: object
      properties:
        coinAmount:
          type: integer
        frequency:
          type: string
          enum: [daily, weekly, monthly]
        isActive:
          type: boolean

    # ── Receipt ───────────────────────────────────────────────────────────────
...