# lib/api-zod/src/generated/types/allowance.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Family Finance & Chores API
 * OpenAPI spec version: 0.1.0
 */
import type { AllowanceFrequency } from './allowanceFrequency';

export interface Allowance {
  id: number;
  familyId: number;
  memberId: number;
  memberName: string;
  coinAmount: number;
  frequency: AllowanceFrequency;
  isActive: boolean;
  /** @nullable */
  nextPayDate?: Date | null;
  createdAt: Date;
}