# lib/api-zod/src/generated/types/allowanceInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Family Finance & Chores API
 * OpenAPI spec version: 0.1.0
 */
import type { AllowanceInputFrequency } from './allowanceInputFrequency';

export interface AllowanceInput {
  memberId: number;
  coinAmount: number;
  frequency: AllowanceInputFrequency;
}