# lib/api-zod/src/generated/types/allowanceUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Family Finance & Chores API
 * OpenAPI spec version: 0.1.0
 */
import type { AllowanceUpdateFrequency } from './allowanceUpdateFrequency';

export interface AllowanceUpdate {
  coinAmount?: number;
  frequency?: AllowanceUpdateFrequency;
  isActive?: boolean;
}