# lib/api-zod/src/generated/types/choreInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Family Finance & Chores API
 * OpenAPI spec version: 0.1.0
 */
import type { ChoreInputRecurringFrequency } from './choreInputRecurringFrequency';

export interface ChoreInput {
  title: string;
  description?: string;
  coinReward: number;
  assigneeId?: number;
  dueDate?: Date;
  isRecurring?: boolean;
  recurringFrequency?: ChoreInputRecurringFrequency;
}